\name{detectionCalculation}\alias{detectionCalculation}\title{calculate p_value of each gene}\description{From the mapping result of specified dataset, calculae the adjusted mean number of reads mapped to a particular GC content in integenic region. This is then treated as background to calculate detection p_value for each gene of same GC content. }\usage{detectionCalculation(dataset)}\arguments{  \item{dataset}{filename from .SAM file}  }\value{ result is given as a density plot on p_value of each gene}%\references{}\author{Zhiyin Dai and Wei Shi}%\note{}%\seealso{}% Add one or more standard keywords, see file 'KEYWORDS' in the % R documentation directory.%\keyword{}%\keyword{}% __ONLY ONE__ keyword per line