\name{detectionCallAnnotation}
\alias{detectionCallAnnotation}
\title{Generate annotation data used for calculating detection p values}
\description{This is for internal use only.}
\usage{
detectionCallAnnotation(species="hg", binsize=2000)
}
\arguments{
  \item{species}{character string specifying the species to analyase}
  \item{binsize}{binsize of integenic region}
  
}
\value{ 
Two files containing annotation information for exons regions and integenic regions, respectively.
}
%\references{}
\author{Zhiyin Dai and Wei Shi}
%\note{}
%\seealso{}
% Add one or more standard keywords, see file 'KEYWORDS' in the 
% R documentation directory.
%\keyword{}
%\keyword{}% __ONLY ONE__ keyword per line

