\name{processExons}
\alias{processExons}
\title{Calculate position of exons from (.md) file}
\description{Calculate position of exons from (.md) file
}
\usage{
processExons(filename="human_seq_gene.md", species="hg")
}
\arguments{
  \item{filename}{ character string giving the name of input .md file}
  \item{species}{ character string specifying the species of given data}
}
\details{
From the given human_seq_gene.md file, we get position of utr and cds
position of exons are then derived for future study
}

\value{ 
An output file is generated  
}
%\references{}
\author{Zhiyin Dai and Wei Shi}
%\note{}
%\seealso{}
% Add one or more standard keywords, see file 'KEYWORDS' in the 
% R documentation directory.
%\keyword{}
%\keyword{}% __ONLY ONE__ keyword per line

