\name{RsubreadUsersGuide}
\alias{RsubreadUsersGuide}

\title{View Rsubread Users Guide}
\description{Users Guide for Rsubread and Subread}

\usage{
RsubreadUsersGuide()
}

%\arguments{}

\value{Character string giving the file location.}

\details{
The Subread/Rsubread Users Guide provides detailed description to the functions and programs included in the Subread and Rsubread software packages.
It also includes case studies for analyzing next-gen sequencing data.

The Subread package is written in C and it can be downloaded from \url{http://subread.sourceforge.net}.
The Rsubread package provides R wrappers functions for many of the programs included in Subread package.

}

\seealso{
\code{\link{vignette}}
}

\author{Wei Shi}

%\examples{}

\keyword{documentation}
