\name{cellCounts}
\alias{cellCounts}
\title{Map and count single-cell RNA data sets}
\description{The \code{cellCounts} function maps and counts reads in single-cell RNA (scRNA) data sets to generate cell-level UMI count tables.}
\usage{
cellCounts(
  # the mandatory arguments
  input.directory,
  index,
  output.BAM,
  sample.sheet,
  cell.barcode.list,

  # the optional arguments related to read mapping
  input.mode="BCL",
  nthreads=16,

  # the optional arguments related to counting
  annot.inbuilt="mm10",
  annot.ext=NULL,
  isGTFAnnotationFile=FALSE,
  GTF.featureType="exon",
  GTF.attrType="gene_id",
  GTF.attrType.extra=NULL,
  chrAliases=NULL,
  useMetaFeatures=TRUE,
  allowMultiOverlap=FALSE,
  countMultiMappingReads=FALSE
) 
}


\arguments{
  \item{input.directory}{character strings giving one or many directories containing the input data sets. Each data set should be put into an individual directory.}
  \item{index}{character string giving the subread index built by the {buildindex} function.}
  \item{output.BAM}{character strings giving one or many output file names, each corrsponding to an input directory. The output files are in the BAM format.}
  \item{sample.sheet}{character strings giving one or many sample sheet files in the CSV format. See the Illumina bcl2fastq manual for details.}
  \item{cell.barcode.list}{character string giving the name of a textual file containing the list of cell barcodes used in sample preparation. The file may be gzipped.}
  \item{input.mode}{character string giving the input mode. Currently only the BCL-format input is supported.}
  \item{nthreads}{numeric value giving the number of threads used for read mapping and counting. 16 by default.}
  \item{annot.inbuilt, annot.ext, isGTFAnnotationFile, GTF.featureType, GTF.attrType, GTF.attrType.extra, chrAliases, useMetaFeatures, allowMultiOverlap, countMultiMappingReads}{options for counting the mapping results. The options have the same meanings as \code{featureCounts}}
}

\details{
  \code{cellCounts} is designed to process single-cell RNA data that is generated by an Illumina sequencer. It does not need the data sets being demultiplexed and converted into fastq files, rather it does demultiplexing and BCL-format decoding on-the-fly. It is capable to process many data sets, each containing many samples, in one run.
}

\value{
  \code{cellCounts} returns a list object containing many elements. The \code{counts.*} elements give the counts of reads to each feature or meta-feature in each sample, the same as the returned object from \code{featureCounts}. The \code{scRNA.table.*} elements are tables giving the feature or meta-feature level UMI counts in each cells. 
}
\author{Wei Shi and Yang Liao}
