\name{detectionCall}\alias{detectionCall}\title{map reads from the given .SAM dataset to exon / integenic region}\description{From the given .SAM file, retrieve read positionthen map this position to see if it belongs to exon or integenic region area}\usage{detectionCall(dataset)}\arguments{  \item{dataset}{filename from .SAM file}  }\value{ Mapping results are given as two text files for exon / integenic region}%\references{}\author{Zhiyin Dai and Wei Shi}%\note{}%\seealso{}% Add one or more standard keywords, see file 'KEYWORDS' in the % R documentation directory.%\keyword{}%\keyword{}% __ONLY ONE__ keyword per line