\name{detectionCallAnnotation}\alias{detectionCallAnnotation}\title{generate binned integenic region file and exon file with GC content}\description{generate binned integenic region file and exon file with GC content}\usage{detectionCallAnnotation(exonfile = "exon.txt", irfile="integenic_region.txt", binsize=2000)}\arguments{  \item{exonfile}{text file containing position about exons}  \item{irfile}{text file containing position about integenic regions}  \item{binsize}{binsize of integenic region}  }\value{ Output file for both exons and integenic region are generated with GC content calculated for eachexon / bin of integenic region}%\references{}\author{Zhiyin Dai and Wei Shi}%\note{}%\seealso{}% Add one or more standard keywords, see file 'KEYWORDS' in the % R documentation directory.%\keyword{}%\keyword{}% __ONLY ONE__ keyword per line